pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-file-rename-n-sort"
version = "26.01"
description = "Local-first macOS organizer that uses metadata-aware plugins and an optional local LLM."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "llm-file-rename-n-sort" }]
dependencies = []

[project.optional-dependencies]
pdf = ["pypdf>=4.0.0"]
docx = ["python-docx>=1.1.0"]
odt = ["odfpy>=1.4.1"]
dev = ["pytest>=7.0.0"]
image = ["pillow>=10.0.0", "pillow-heif>=0.18.0"]

[project.scripts]
llm-file-rename-n-sort = "rename_n_sort.cli:main"
